requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
celery-progress==0.1.1
certifi==2021.5.30
charset-normalizer==2.0.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-celery==3.3.1
django-celery-results==2.2.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
lxml==4.6.3
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
prompt-toolkit==3.0.19
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
